rent_first/requirements.txt
pytidylib == 0.2.1
requests == 1.2.3
BeautifulSoup == 4.2.1
BeautifulSoup == 3.2.1